+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Sort attribute values in tiles to make them compress a little better
 * Fix dropping of "largest" points when there are duplicate points
 * Add an option to prevent guessing a basezoom higher than the maxzoom
+* Add --order-by and --order-descending options
 
 ## 2.4.1